charge protect: change flip to 10s throttle pulse
Cannot find where the max throttle is set, so uncertain if the thrust can be increased
At the same time, I figure that a 10s pulse at high level should be able to reduce the
battery voltage to a safer level. Ideally then that should give the baseboard time to
realise that it's meant to cut to trickle charge or disable charging.